+++ .gitignore
@@ -2,3 +2,4 @@ node_modules
 dist/
 ci-work/
 coverage/
+.vscode
